+++ lib/levels/bumpers/bumpers-one.js
@@ -1,11 +1,19 @@
 var Bumper = require('../../bumper')
 
-var bumpersOne = [{minX: 50, minY: 50, maxX: 250, maxY: 60},
-  {minX: 50, minY: 450, maxX: 250, maxY: 460},
-  {minX: 50, minY: 50, maxX: 60, maxY: 450},
-  {minX: 250, minY: 50, maxX: 260, maxY: 460},
-  {minX: 90, minY: 150, maxX: 210, maxY: 160, color: "blue"}
+var bumpersOne = [{minX: 50, minY: 50, maxX: 660, maxY: 60},
+  {minX: 50, minY: 550, maxX: 650, maxY: 560},
+  {minX: 50, minY: 50, maxX: 60, maxY: 550},
+  {minX: 650, minY: 50, maxX: 660, maxY: 560},
+  // {minX: 90, minY: 150, maxX: 210, maxY: 160, color: "blue"}
   ]
+
+
+// var bumpersOne = [{minX: 50, minY: 50, maxX: 250, maxY: 60},
+//   {minX: 50, minY: 450, maxX: 250, maxY: 460},
+//   {minX: 50, minY: 50, maxX: 60, maxY: 450},
+//   {minX: 250, minY: 50, maxX: 260, maxY: 460},
+//   {minX: 90, minY: 150, maxX: 210, maxY: 160, color: "blue"}
+//   ]
 
 
 module.exports = bumpersOne;
