+++ lib/levels/bumpers/bumpers-three.js
@@ -1,0 +1,11 @@
+var Bumper = require('../../bumper')
+
+var bumpersOne = [{minX: 50, minY: 50, maxX: 250, maxY: 60},
+  {minX: 50, minY: 450, maxX: 250, maxY: 460},
+  {minX: 50, minY: 50, maxX: 60, maxY: 450},
+  {minX: 250, minY: 50, maxX: 260, maxY: 460},
+  {minX: 90, minY: 150, maxX: 210, maxY: 160, color: "blue"}
+  ]
+
+
+module.exports = bumpersOne;
